Add files via upload
problem Statement:
Prediction of Gold Prices Using a 10-Year Dataset

Description:
We aim to predict gold prices by training a machine learning model on a dataset that spans several days over a 10-year period.

Data Overview:
The dataset is in CSV format, comprising 2,290 rows and 7 columns. It includes 5 numerical columns and one date column. The variables SPX, GLD, USO, SLV, and EUR/USD are recorded against the corresponding dates.

Model Building and Training:

This is a regression problem.
We will use a Random Forest Regressor to build our model.
The model will be trained on the training data and then tested with sample test data to predict gold prices.